refactor(HW:FMEA): move methods to FMEA class
Load everything into the big (D)FME(C)A treeview and then hide the
columns not used for the Functional FMEA.  Hidden columns will just be
loaded with default values.  This simplifies the FMEA view.
* Move __load_combobox() method from FFMEA and DFMEA classes to FMEA
class.
* Move __set_callbacks() method from FFMEA and DFMEA classes to FMEA
class.
* Move _do_change_row() method from FFMEA and DFMEA classes to FMEA
class.
* Move _do_load_page() method from FFMEA and DFMEA classes to FMEA
class.
* Move _do_request_insert() from FFMEA and DFMEA classes to FMEA class.
* Move _on_cell_edit() from FFMEA and DFMEA classes to FMEA class.
* Remove FFMEA and DFMEA layout file and replace with FMEA.xml file.
